+++ bindings/dart/dart/lib/src/native_bindings.dart
@@ -213,6 +213,15 @@ typedef _DbStringOutDart = int Function(
   Pointer<Pointer<Utf8>> outValue,
 );
 
+typedef _DbInspectStorageStateC = Uint32 Function(
+  Pointer<DdbDb> db,
+  Pointer<Pointer<Utf8>> outJson,
+);
+typedef _DbInspectStorageStateDart = int Function(
+  Pointer<DdbDb> db,
+  Pointer<Pointer<Utf8>> outJson,
+);
+
 typedef _DbNamedStringOutC = Uint32 Function(
   Pointer<DdbDb> db,
   Pointer<Utf8> name,
@@ -696,6 +705,9 @@ class NativeBindings {
         dbGetSchemaSnapshotJson =
             _lib.lookupFunction<_DbStringOutC, _DbStringOutDart>(
                 'ddb_db_get_schema_snapshot_json'),
+        dbInspectStorageStateJson = _lib.lookupFunction<
+                _DbInspectStorageStateC, _DbInspectStorageStateDart>(
+            'ddb_db_inspect_storage_state_json'),
         evictSharedWal =
             _lib.lookupFunction<_EvictSharedWalC, _EvictSharedWalDart>(
                 'ddb_evict_shared_wal'),
@@ -825,6 +837,7 @@ class NativeBindings {
   final _DbNamedStringOutDart dbGetViewDdl;
   final _DbStringOutDart dbListTriggersJson;
   final _DbStringOutDart dbGetSchemaSnapshotJson;
+  final _DbInspectStorageStateDart dbInspectStorageStateJson;
   final _EvictSharedWalDart evictSharedWal;
 
   // ddb_db_execute result accessors